reorg the reorg, refactor types a bit
