+++ client/scripts/views/components/component_kit/types.ts
@@ -1,17 +1,28 @@
+export type State = Record<string, unknown>;
+
 export enum ComponentType {
   Button = 'Button',
   GradientButton = 'GradientButton',
 }
+
+export type StyleProps = {
+  disabled?: boolean;
+  className?: string;
+};
 
 export enum ButtonType {
   Primary = 'primary',
   Secondary = 'secondary',
 }
 
-export type StyleProps = {
+export type ButtonStyleProps = {
   buttonType: ButtonType;
-  disabled?: boolean;
-  className?: string;
+} & StyleProps;
+
+export type ButtonProps = {
+  styleProps: ButtonStyleProps;
+  label: string;
+  onclick: (e?: MouseEvent) => void;
 };
 
 export enum Justify {
